externalised the router
