--- frontend/WebAppHarmony/Routes/Router.js
@@ -1,2 +1,0 @@
-export const MODERATEPAIRING = "/ModeratePairing";
-export const HOME = "/App";
